requirements.txt
numpy~=1.24.2
pandas
matplotlib>=3.7.1
scikit-learn
Pillow~=9.4.0
tqdm>=4.65.0

torch>=2.0.0+cu117
torchvision>=0.15.1+cu117
opencv-python>=4.7.0.72
scipy>=1.10.1